requirements.txt
onnxruntime
tokenizers>=0.13.2
numpy
opencv-python
Pillow>=9.2.0
PyYAML
cython